Remove unused crate
