+++ .github/workflows/mutants.yml
@@ -86,10 +86,8 @@ jobs:
       - name: Establish baseline build
         # Workspace-level `--features` only activates the named features
         # on packages that declare them, so this is safe across the
-        # whole workspace. `rayon` is declared by `automerge_sedimentree`
-        # and gates its parallel ingest variants + the corresponding
-        # equivalence test.
-        run: cargo build --workspace --tests --features arbitrary,metrics,rayon
+        # whole workspace.
+        run: cargo build --workspace --tests --features arbitrary,metrics
 
       - name: Compute diff vs base
         id: diff
@@ -124,7 +122,6 @@ jobs:
             [subduction_iroh]=""
             [subduction_keyhive]=""
             [subduction_websocket]=""
-            [automerge_sedimentree]="rayon"
           )
 
           # Extract crate names from changed file paths. Top-level dirs
@@ -234,15 +231,6 @@ jobs:
             features: ""
           - crate: subduction_websocket
             features: ""
-          - crate: automerge_sedimentree
-            # `rayon` activates the per-fragment / chunked parallel ingest
-            # variants. They live behind `#[cfg(feature = "rayon")]` and
-            # their only test caller (`parallel_variants_match_sequential`
-            # in `tests/real_world_ingestion.rs`) is similarly gated.
-            # Without this flag those functions compile to no-ops in the
-            # mutant baseline and the `-> Default::default()` mutations
-            # against them go silently uncaught.
-            features: "rayon"
 
     permissions:
       contents: read
