--- automerge_sedimentree/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-name = "automerge_sedimentree"
-version = "0.10.1"
-description = "Adapter for storing Automerge documents in Sedimentree"
-
-categories = ["web-programming"]
-keywords = ["automerge", "sedimentree"]
-readme = "./README.md"
-
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-automerge = { workspace = true }
-rayon = { workspace = true, optional = true }
-sedimentree_core = { workspace = true }
-
-[dev-dependencies]
-automerge_sedimentree = { path = ".", features = ["std"] }
-criterion = { workspace = true }
-criterion_pprof = { workspace = true }
-rand = { workspace = true, features = ["small_rng"] }
-testresult = { workspace = true }
-
-[[bench]]
-name = "egwalker"
-harness = false
-
-[features]
-default = []
-rayon = ["dep:rayon", "sedimentree_core/rayon", "std"]
-std = ["sedimentree_core/std"]
-
-[lints]
-workspace = true
