--- automerge_sedimentree/README.md
@@ -1,5 +1,0 @@
-# Subduction Automerge
-## Automerge integration for Subduction.
-
-🚧 This is an early release preview. It has a very unstable API. No guarantees are given. DO NOT use for production use cases at this time. USE AT YOUR OWN RISK. 🚧
-
