--- automerge_sedimentree/tests/real_world_ingestion.rs
@@ -1,1152 +1,0 @@
-//! Roundtrip ingestion tests using real Automerge documents from the egwalker
-//! paper test vectors.
-//!
-//! Each test loads a `.am` file, decomposes it via [`ingest::ingest_automerge`]
-//! (which delegates fragmentation to upstream [`Automerge::fragments`] +
-//! [`Automerge::bundle_fragments`]) and verifies structural invariants. A
-//! couple of helpers in this file still use the legacy
-//! `IndexedSedimentreeAutomerge` + `build_fragment_store` path as a
-//! reference; those are explicitly noted at their call sites. Byte-identical
-//! reassembly is tested in release mode only (automerge's `get_changes`
-//! has a `debug_assert` that doubles work).
-//!
-//! [`ingest::ingest_automerge`]: automerge_sedimentree::ingest::ingest_automerge
-//! [`Automerge::fragments`]: automerge::Automerge::fragments
-//! [`Automerge::bundle_fragments`]: automerge::Automerge::bundle_fragments
-
-#![allow(
-    clippy::cast_precision_loss,
-    clippy::expect_used,
-    clippy::indexing_slicing,
-    clippy::panic,
-    clippy::unwrap_used
-)]
-
-use automerge::{Automerge, ChangeHash, ROOT, ReadDoc};
-use automerge_sedimentree::indexed::{IndexedSedimentreeAutomerge, OwnedParents};
-use sedimentree_core::{
-    blob::{Blob, BlobMeta},
-    collections::{Map, Set},
-    commit::{CommitStore, CountLeadingZeroBytes, FragmentState},
-    crypto::{digest::Digest, fingerprint::FingerprintSeed},
-    fragment::Fragment,
-    id::SedimentreeId,
-    loose_commit::{LooseCommit, id::CommitId},
-    sedimentree::{Sedimentree, SedimentreeItem},
-};
-
-/// Lightweight decomposition using only `get_changes_meta` (fast in debug).
-/// Does NOT extract raw bytes — just builds the fragment index.
-struct MetadataDecomp {
-    change_count: usize,
-    heads: Vec<CommitId>,
-    covered: Set<CommitId>,
-    uncovered_count: usize,
-    fragment_count: usize,
-}
-
-fn decompose_meta(doc: &Automerge) -> MetadataDecomp {
-    let metadata = doc.get_changes_meta(&[]);
-    let change_count = metadata.len();
-    let all_digests: Set<CommitId> = metadata.iter().map(|m| CommitId::new(m.hash.0)).collect();
-
-    let store = IndexedSedimentreeAutomerge::from_metadata(&metadata);
-    let heads: Vec<CommitId> = doc.get_heads().iter().map(|h| CommitId::new(h.0)).collect();
-    let mut known: Map<CommitId, FragmentState<OwnedParents>> = Map::new();
-
-    let fresh = store
-        .build_fragment_store(&heads, &mut known, &CountLeadingZeroBytes)
-        .expect("build_fragment_store");
-    let fragment_count = fresh.len();
-
-    let covered: Set<CommitId> = known
-        .values()
-        .flat_map(|s| s.members().iter().copied())
-        .collect();
-
-    let uncovered_count = all_digests.iter().filter(|d| !covered.contains(d)).count();
-
-    MetadataDecomp {
-        change_count,
-        heads,
-        covered,
-        uncovered_count,
-        fragment_count,
-    }
-}
-
-/// Full decomposition including raw bytes (requires `get_changes` — slow in
-/// debug due to automerge's internal `debug_assert`).
-struct FullDecomp {
-    change_count: usize,
-    /// Compact Document-format blobs per fragment.
-    fragment_blobs: Vec<Vec<u8>>,
-    /// Raw change bytes per loose commit.
-    uncovered_blobs: Vec<Vec<u8>>,
-    uncovered_parents: Vec<std::collections::BTreeSet<CommitId>>,
-    fragment_state_blobs: Vec<(FragmentState<OwnedParents>, Blob)>,
-}
-
-fn decompose_full(doc: &Automerge) -> FullDecomp {
-    let metadata = doc.get_changes_meta(&[]);
-    let change_count = metadata.len();
-
-    let store = IndexedSedimentreeAutomerge::from_metadata(&metadata);
-    let heads: Vec<CommitId> = doc.get_heads().iter().map(|h| CommitId::new(h.0)).collect();
-    let mut known: Map<CommitId, FragmentState<OwnedParents>> = Map::new();
-
-    let fresh = store
-        .build_fragment_store(&heads, &mut known, &CountLeadingZeroBytes)
-        .expect("build_fragment_store");
-    let states: Vec<_> = fresh.into_iter().cloned().collect();
-
-    let covered: Set<CommitId> = known
-        .values()
-        .flat_map(|s| s.members().iter().copied())
-        .collect();
-
-    // Get all changes with raw bytes in topological order.
-    let changes = doc.get_changes(&[]);
-
-    // For each fragment: concatenate only the member changes' raw bytes,
-    // preserving the topological order from get_changes.
-    let mut fragment_blobs = Vec::new();
-    let mut fragment_state_blobs = Vec::new();
-    for state in &states {
-        let members: Set<ChangeHash> = state
-            .members()
-            .iter()
-            .map(|d| ChangeHash(*d.as_bytes()))
-            .collect();
-
-        let mut raw = Vec::new();
-        for change in &changes {
-            if members.contains(&change.hash()) {
-                raw.extend_from_slice(change.raw_bytes());
-            }
-        }
-
-        fragment_state_blobs.push((state.clone(), Blob::new(raw.clone())));
-        fragment_blobs.push(raw);
-    }
-
-    // Loose commits: raw_bytes per uncovered change.
-    let mut uncovered_blobs = Vec::new();
-    let mut uncovered_parents = Vec::new();
-    for change in &changes {
-        let id = CommitId::new(change.hash().0);
-        if covered.contains(&id) {
-            continue;
-        }
-        uncovered_blobs.push(change.raw_bytes().to_vec());
-        uncovered_parents.push(change.deps().iter().map(|d| CommitId::new(d.0)).collect());
-    }
-
-    FullDecomp {
-        change_count,
-        fragment_blobs,
-        uncovered_blobs,
-        uncovered_parents,
-        fragment_state_blobs,
-    }
-}
-
-fn sed_id(bytes: &[u8]) -> SedimentreeId {
-    let d: Digest<Blob> = Digest::hash(&Blob::new(bytes.to_vec()));
-    SedimentreeId::new(*d.as_bytes())
-}
-
-fn build_tree(bytes: &[u8], d: &FullDecomp) -> (Sedimentree, Vec<Fragment>, Vec<LooseCommit>) {
-    let id = sed_id(bytes);
-
-    let fragments: Vec<Fragment> = d
-        .fragment_state_blobs
-        .iter()
-        .map(|(state, blob)| state.clone().to_fragment(id, BlobMeta::new(blob)))
-        .collect();
-
-    let loose: Vec<LooseCommit> = d
-        .uncovered_blobs
-        .iter()
-        .zip(d.uncovered_parents.iter())
-        .enumerate()
-        .map(|(i, (raw, parents))| {
-            let head = CommitId::new({
-                let mut bytes = [0u8; 32];
-                #[allow(clippy::cast_possible_truncation)]
-                {
-                    bytes[0] = i as u8;
-                    bytes[1] = (i >> 8) as u8;
-                }
-                // Use raw bytes hash for uniqueness
-                let blob = Blob::new(raw.clone());
-                let digest = Digest::hash(&blob);
-                bytes[2..].copy_from_slice(&digest.as_bytes()[..30]);
-                bytes
-            });
-            LooseCommit::new(
-                id,
-                head,
-                parents.clone(),
-                BlobMeta::new(&Blob::new(raw.clone())),
-            )
-        })
-        .collect();
-
-    let tree = Sedimentree::new(fragments.clone(), loose.clone());
-    (tree, fragments, loose)
-}
-
-// ---------------------------------------------------------------------------
-// Fast tests (metadata only — no get_changes, runs in debug mode)
-// ---------------------------------------------------------------------------
-
-/// Verify documents load and `build_fragment_store` completes.
-#[test]
-fn load_and_count() {
-    for (name, bytes) in [
-        ("C1", &include_bytes!("../test-vectors/C1.am")[..]),
-        ("C2", &include_bytes!("../test-vectors/C2.am")[..]),
-        ("S1", &include_bytes!("../test-vectors/S1.am")[..]),
-        ("S2", &include_bytes!("../test-vectors/S2.am")[..]),
-        ("S3", &include_bytes!("../test-vectors/S3.am")[..]),
-        ("A1", &include_bytes!("../test-vectors/A1.am")[..]),
-        ("A2", &include_bytes!("../test-vectors/A2.am")[..]),
-    ] {
-        let doc = Automerge::load(bytes).expect(name);
-        let d = decompose_meta(&doc);
-        eprintln!(
-            "{name}: {} changes, {} fragments, {} uncovered, {} heads",
-            d.change_count,
-            d.fragment_count,
-            d.uncovered_count,
-            d.heads.len(),
-        );
-        assert!(d.change_count > 0, "{name}: should have changes");
-        assert!(!d.heads.is_empty(), "{name}: should have heads");
-    }
-}
-
-/// Fragment count + uncovered count == total change count.
-#[test]
-fn coverage_is_complete() {
-    for (name, bytes) in [
-        ("C1", &include_bytes!("../test-vectors/C1.am")[..]),
-        ("C2", &include_bytes!("../test-vectors/C2.am")[..]),
-        ("S1", &include_bytes!("../test-vectors/S1.am")[..]),
-        ("S2", &include_bytes!("../test-vectors/S2.am")[..]),
-        ("S3", &include_bytes!("../test-vectors/S3.am")[..]),
-        ("A1", &include_bytes!("../test-vectors/A1.am")[..]),
-        ("A2", &include_bytes!("../test-vectors/A2.am")[..]),
-    ] {
-        let doc = Automerge::load(bytes).expect(name);
-        let d = decompose_meta(&doc);
-        let total = d.covered.len() + d.uncovered_count;
-        assert_eq!(
-            total,
-            d.change_count,
-            "{name}: covered ({}) + uncovered ({}) != total ({})",
-            d.covered.len(),
-            d.uncovered_count,
-            d.change_count,
-        );
-    }
-}
-
-/// Documents with >2 changes should produce at least one fragment.
-#[test]
-fn produces_fragments() {
-    for (name, bytes) in [
-        ("A1", &include_bytes!("../test-vectors/A1.am")[..]),
-        ("A2", &include_bytes!("../test-vectors/A2.am")[..]),
-        // C1/C2 have 93k/134k changes — always produce fragments
-        ("C1", &include_bytes!("../test-vectors/C1.am")[..]),
-        ("C2", &include_bytes!("../test-vectors/C2.am")[..]),
-    ] {
-        let doc = Automerge::load(bytes).expect(name);
-        let d = decompose_meta(&doc);
-        assert!(
-            d.fragment_count > 0,
-            "{name}: {} changes should produce fragments",
-            d.change_count,
-        );
-    }
-}
-
-fn roundtrip_full(name: &str, bytes: &[u8]) {
-    let doc = Automerge::load(bytes).expect(name);
-    let d = decompose_full(&doc);
-    let (tree, _fragments, _loose) = build_tree(bytes, &d);
-
-    // Build a lookup from BlobMeta digest → raw bytes.
-    // Sedimentree::fragments()/loose_commits() iterate in BTreeMap order
-    // (by Digest<Fragment>/Digest<LooseCommit>), which differs from the
-    // FullDecomp insertion order. We match them via BlobMeta digest.
-    // Map blob digest → raw bytes for lookup during reassembly.
-    let blob_by_digest: Map<Digest<Blob>, &[u8]> = d
-        .fragment_state_blobs
-        .iter()
-        .map(|(_, blob)| (Digest::hash(blob), blob.as_slice()))
-        .chain(d.uncovered_blobs.iter().map(|raw| {
-            let blob = Blob::new(raw.clone());
-            (Digest::hash(&blob), raw.as_slice())
-        }))
-        .collect();
-
-    let order = tree
-        .topsorted_blob_order()
-        .expect("no cycles in test fixture");
-    let fragments: Vec<_> = tree.fragments().collect();
-    let loose: Vec<_> = tree.loose_commits().collect();
-
-    let mut buf = Vec::new();
-    for blob_ref in &order {
-        let digest = match blob_ref {
-            SedimentreeItem::Fragment(i) => fragments[*i].summary().blob_meta().digest(),
-            SedimentreeItem::LooseCommit(i) => loose[*i].blob_meta().digest(),
-        };
-        let raw = blob_by_digest
-            .get(&digest)
-            .unwrap_or_else(|| panic!("{name}: blob not found for {blob_ref:?}"));
-        buf.extend_from_slice(raw);
-    }
-
-    let mut rebuilt = Automerge::new();
-    rebuilt
-        .load_incremental(&buf)
-        .expect("load topsorted blobs");
-
-    assert_eq!(
-        doc.get_heads(),
-        rebuilt.get_heads(),
-        "{name}: heads diverged"
-    );
-    assert_eq!(
-        d.change_count,
-        rebuilt.get_changes_meta(&[]).len(),
-        "{name}: change count diverged"
-    );
-
-    // Content verification: compare root-level keys and values.
-    let orig_keys: Vec<_> = doc.keys(&ROOT).collect();
-    let rebuilt_keys: Vec<_> = rebuilt.keys(&ROOT).collect();
-    assert_eq!(orig_keys, rebuilt_keys, "{name}: root keys diverged");
-
-    for key in &orig_keys {
-        let orig_val = doc.get(&ROOT, key.as_str());
-        let rebuilt_val = rebuilt.get(&ROOT, key.as_str());
-        assert_eq!(
-            orig_val, rebuilt_val,
-            "{name}: value at root key {key:?} diverged"
-        );
-    }
-
-    // Spot check: root object length should match.
-    assert_eq!(
-        doc.length(&ROOT),
-        rebuilt.length(&ROOT),
-        "{name}: root object length diverged"
-    );
-}
-
-/// Report blob sizes for all vectors that have fragments.
-/// Release-only because it calls `get_changes`.
-#[test]
-#[cfg_attr(debug_assertions, ignore = "too slow in debug mode")]
-fn blob_size_report() {
-    for (name, bytes) in [
-        ("A1", &include_bytes!("../test-vectors/A1.am")[..]),
-        ("A2", &include_bytes!("../test-vectors/A2.am")[..]),
-        ("C1", &include_bytes!("../test-vectors/C1.am")[..]),
-        ("C2", &include_bytes!("../test-vectors/C2.am")[..]),
-    ] {
-        let doc = Automerge::load(bytes).expect(name);
-        let d = decompose_full(&doc);
-        let frag_bytes: usize = d.fragment_blobs.iter().map(Vec::len).sum();
-        let loose_bytes: usize = d.uncovered_blobs.iter().map(Vec::len).sum();
-        let original = bytes.len();
-        let ratio = (frag_bytes + loose_bytes) as f64 / original as f64;
-        eprintln!(
-            "{name}: original={original} fragments={frag_bytes} ({} blobs) loose={loose_bytes} ({} blobs) total={} ratio={ratio:.2}x",
-            d.fragment_state_blobs.len(),
-            d.uncovered_blobs.len(),
-            frag_bytes + loose_bytes,
-        );
-    }
-}
-
-#[test]
-fn roundtrip_s1() {
-    roundtrip_full("S1", include_bytes!("../test-vectors/S1.am"));
-}
-
-#[test]
-fn roundtrip_s2() {
-    roundtrip_full("S2", include_bytes!("../test-vectors/S2.am"));
-}
-
-#[test]
-fn roundtrip_s3() {
-    roundtrip_full("S3", include_bytes!("../test-vectors/S3.am"));
-}
-
-// A1/A2/C1/C2 full roundtrips — only run in release mode because
-// automerge's get_changes has a debug_assert that doubles work.
-#[test]
-#[cfg_attr(debug_assertions, ignore = "too slow in debug mode")]
-fn roundtrip_a1() {
-    roundtrip_full("A1", include_bytes!("../test-vectors/A1.am"));
-}
-
-#[test]
-#[cfg_attr(debug_assertions, ignore = "too slow in debug mode")]
-fn roundtrip_a2() {
-    roundtrip_full("A2", include_bytes!("../test-vectors/A2.am"));
-}
-
-#[test]
-#[cfg_attr(debug_assertions, ignore = "too slow in debug mode")]
-fn roundtrip_c1() {
-    roundtrip_full("C1", include_bytes!("../test-vectors/C1.am"));
-}
-
-#[test]
-#[cfg_attr(debug_assertions, ignore = "too slow in debug mode")]
-fn roundtrip_c2() {
-    roundtrip_full("C2", include_bytes!("../test-vectors/C2.am"));
-}
-
-static S1: &[u8] = include_bytes!("../test-vectors/S1.am");
-
-#[test]
-fn sedimentree_diff_against_empty() {
-    let doc = Automerge::load(S1).unwrap();
-    let d = decompose_full(&doc);
-    let (tree, fragments, loose) = build_tree(S1, &d);
-    let empty = Sedimentree::default();
-
-    let diff = empty.diff(&tree);
-    assert_eq!(diff.left_missing_fragments.len(), fragments.len());
-    assert_eq!(diff.left_missing_commits.len(), loose.len());
-    assert!(diff.right_missing_fragments.is_empty());
-    assert!(diff.right_missing_commits.is_empty());
-}
-
-#[test]
-fn sedimentree_diff_against_self() {
-    let doc = Automerge::load(S1).unwrap();
-    let d = decompose_full(&doc);
-    let (tree, _, _) = build_tree(S1, &d);
-
-    let diff = tree.diff(&tree);
-    assert!(diff.left_missing_fragments.is_empty());
-    assert!(diff.left_missing_commits.is_empty());
-    assert!(diff.right_missing_fragments.is_empty());
-    assert!(diff.right_missing_commits.is_empty());
-}
-
-#[test]
-fn fingerprint_self_diff_empty() {
-    let doc = Automerge::load(S1).unwrap();
-    let d = decompose_full(&doc);
-    let (tree, _, _) = build_tree(S1, &d);
-
-    let seed = FingerprintSeed::new(42, 99);
-    let summary = tree.fingerprint_summarize(&seed);
-    let diff = tree.diff_remote_fingerprints(&summary);
-
-    assert!(diff.local_only_fragments.is_empty());
-    assert!(diff.local_only_commits.is_empty());
-    assert!(diff.remote_only_fragment_fingerprints.is_empty());
-    assert!(diff.remote_only_commit_fingerprints.is_empty());
-}
-
-#[test]
-fn merge_identical_is_idempotent() {
-    let doc = Automerge::load(S1).unwrap();
-    let d = decompose_full(&doc);
-    let (tree_a, frags, loose) = build_tree(S1, &d);
-    let tree_b = Sedimentree::new(frags, loose);
-    let m = CountLeadingZeroBytes;
-
-    let before = tree_a.minimal_hash(&m);
-    let mut merged = tree_a;
-    merged.merge(tree_b);
-    assert_eq!(before, merged.minimal_hash(&m));
-}
-
-/// After minimize, the sedimentree should still have the same number of
-/// loose commits and fragments as before. Loose commits' parents are
-/// remapped to fragment heads so minimize doesn't prune them.
-#[test]
-fn minimize_preserves_all_items_after_ingestion() {
-    let doc = Automerge::load(include_bytes!("../test-vectors/S1.am")).unwrap();
-    let sed_id = sed_id(include_bytes!("../test-vectors/S1.am"));
-    let result = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-
-    let before_fragments = result.sedimentree.fragments().count();
-    let before_commits = result.sedimentree.loose_commits().count();
-
-    let minimized = result.sedimentree.minimize(&CountLeadingZeroBytes);
-    let after_fragments = minimized.fragments().count();
-    let after_commits = minimized.loose_commits().count();
-
-    assert_eq!(
-        before_commits, after_commits,
-        "minimize should not drop loose commits (before={before_commits}, after={after_commits})"
-    );
-    assert!(
-        after_fragments <= before_fragments,
-        "minimize should not add fragments"
-    );
-}
-
-/// Every parent of every loose commit produced by [`ingest_automerge`] must be
-/// either a fragment head or another loose-commit head in the same
-/// sedimentree. If a parent ever references an _interior_ member of a cached
-/// fragment, the `member_to_head` remap in `finalize` is broken — and
-/// [`Sedimentree::minimize`] will subsequently mistake that loose commit for
-/// covered data and prune it.
-///
-/// Guards the `id != head` filter at `ingest.rs:189` against mutation. With
-/// `!=` flipped to `==`, the map ends up containing only `head → head`
-/// (identity) entries, so interior members are never remapped to their
-/// containing fragment's head and instead leak into loose-commit parent
-/// sets. A1 is the smallest egwalker vector with both cached fragments and
-/// loose commits whose boundary touches interior fragment members, so it
-/// reliably exercises the remap.
-///
-/// [`ingest_automerge`]: automerge_sedimentree::ingest::ingest_automerge
-/// [`Sedimentree::minimize`]: sedimentree_core::sedimentree::Sedimentree::minimize
-#[test]
-fn loose_commit_parents_only_reference_known_heads() {
-    let bytes = include_bytes!("../test-vectors/A1.am");
-    let doc = Automerge::load(bytes).expect("A1");
-    let sed_id = sed_id(bytes);
-    let result = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-
-    let fragment_heads: Set<CommitId> =
-        result.sedimentree.fragments().map(Fragment::head).collect();
-    let loose_heads: Set<CommitId> = result
-        .sedimentree
-        .loose_commits()
-        .map(LooseCommit::head)
-        .collect();
-
-    assert!(
-        !fragment_heads.is_empty(),
-        "A1 must produce cached fragments for this test to exercise the remap"
-    );
-    assert!(
-        !loose_heads.is_empty(),
-        "A1 must produce loose commits for this test to exercise the remap"
-    );
-
-    let mut unknown: Vec<(CommitId, CommitId)> = Vec::new();
-    for loose in result.sedimentree.loose_commits() {
-        for parent in loose.parents() {
-            if !fragment_heads.contains(parent)
-                && !loose_heads.contains(parent)
-                && unknown.len() < 5
-            {
-                unknown.push((loose.head(), *parent));
-            }
-        }
-    }
-
-    assert!(
-        unknown.is_empty(),
-        "A1: loose-commit parents reference {} commit(s) that are neither fragment heads \
-         nor loose-commit heads (sample: {unknown:?}). \
-         The member_to_head remap in `finalize` (ingest.rs:189) is broken.",
-        unknown.len(),
-    );
-}
-
-/// The three ingest variants — sequential, per-fragment parallel, and
-/// chunked-parallel — must agree on which commits ended up where (counts,
-/// fragment heads, loose-commit heads, loose-commit parent sets). Blob
-/// byte representation can legitimately differ between variants because
-/// `bundle_fragments` may emit slightly different per-fragment layouts
-/// depending on how many fragments are bundled in a single call — that
-/// surface lives upstream and is exercised by upstream tests.
-///
-/// Guards `ingest_automerge_par` and `ingest_automerge_par_chunked` against
-/// being silently mutated to `Default::default()` — the empty `IngestResult`
-/// would trivially fail the non-empty smoke check below and the count
-/// equalities. A1 (~772 KiB, 7 cached fragments, 184 loose commits) is
-/// the smallest egwalker vector that exercises both code paths.
-#[cfg(feature = "rayon")]
-#[test]
-fn parallel_variants_match_sequential() {
-    use std::collections::BTreeSet;
-
-    use automerge_sedimentree::ingest::{
-        ingest_automerge, ingest_automerge_par, ingest_automerge_par_chunked,
-    };
-
-    let bytes = include_bytes!("../test-vectors/A1.am");
-    let doc = Automerge::load(bytes).expect("A1");
-    let sed_id = sed_id(bytes);
-
-    let seq = ingest_automerge(&doc, sed_id);
-    let par = ingest_automerge_par(&doc, sed_id);
-    let chk = ingest_automerge_par_chunked(&doc, sed_id);
-
-    // Non-empty smoke check: catches `-> Default::default()` mutations that
-    // would otherwise reduce every assertion below to a vacuous `0 == 0`.
-    assert!(seq.fragment_count > 0, "A1 must produce cached fragments");
-    assert!(seq.loose_count > 0, "A1 must produce loose commits");
-    assert!(seq.change_count > 0, "A1 must contain changes");
-    assert!(par.fragment_count > 0, "par result must not be empty");
-    assert!(par.loose_count > 0, "par result must not be empty");
-    assert!(chk.fragment_count > 0, "chunked result must not be empty");
-    assert!(chk.loose_count > 0, "chunked result must not be empty");
-
-    // Counts must agree across variants.
-    assert_eq!(seq.fragment_count, par.fragment_count, "par fragment_count");
-    assert_eq!(
-        seq.fragment_count, chk.fragment_count,
-        "chunked fragment_count"
-    );
-    assert_eq!(seq.loose_count, par.loose_count, "par loose_count");
-    assert_eq!(seq.loose_count, chk.loose_count, "chunked loose_count");
-    assert_eq!(seq.change_count, par.change_count, "par change_count");
-    assert_eq!(seq.change_count, chk.change_count, "chunked change_count");
-    assert_eq!(seq.covered_count, par.covered_count, "par covered_count");
-    assert_eq!(
-        seq.covered_count, chk.covered_count,
-        "chunked covered_count"
-    );
-
-    // Logical structure: the set of fragment heads, the set of loose-commit
-    // heads, and each loose commit's parent set must be identical. These
-    // are content-addressed identities, so they're insensitive to the
-    // per-fragment blob byte differences described above.
-    let fragment_heads = |r: &automerge_sedimentree::ingest::IngestResult| -> BTreeSet<CommitId> {
-        r.sedimentree.fragments().map(Fragment::head).collect()
-    };
-    let loose_heads = |r: &automerge_sedimentree::ingest::IngestResult| -> BTreeSet<CommitId> {
-        r.sedimentree
-            .loose_commits()
-            .map(LooseCommit::head)
-            .collect()
-    };
-    let loose_parents =
-        |r: &automerge_sedimentree::ingest::IngestResult| -> Map<CommitId, BTreeSet<CommitId>> {
-            r.sedimentree
-                .loose_commits()
-                .map(|c| (c.head(), c.parents().clone()))
-                .collect()
-        };
-
-    assert_eq!(
-        fragment_heads(&seq),
-        fragment_heads(&par),
-        "par fragment heads diverged"
-    );
-    assert_eq!(
-        fragment_heads(&seq),
-        fragment_heads(&chk),
-        "chunked fragment heads diverged"
-    );
-    assert_eq!(
-        loose_heads(&seq),
-        loose_heads(&par),
-        "par loose-commit heads diverged"
-    );
-    assert_eq!(
-        loose_heads(&seq),
-        loose_heads(&chk),
-        "chunked loose-commit heads diverged"
-    );
-    assert_eq!(
-        loose_parents(&seq),
-        loose_parents(&par),
-        "par loose parents diverged"
-    );
-    assert_eq!(
-        loose_parents(&seq),
-        loose_parents(&chk),
-        "chunked loose parents diverged"
-    );
-}
-
-/// Fingerprint summary after ingestion should include all items.
-#[test]
-fn fingerprint_summary_includes_all_items_after_ingestion() {
-    let doc = Automerge::load(include_bytes!("../test-vectors/S1.am")).unwrap();
-    let sed_id = sed_id(include_bytes!("../test-vectors/S1.am"));
-    let result = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-
-    let seed = FingerprintSeed::new(42, 99);
-    let summary = result.sedimentree.fingerprint_summarize(&seed);
-
-    assert_eq!(
-        summary.commit_fingerprints().len(),
-        result.loose_count,
-        "fingerprint summary should have all loose commits"
-    );
-    assert_eq!(
-        summary.fragment_fingerprints().len(),
-        result.fragment_count,
-        "fingerprint summary should have all fragments"
-    );
-}
-
-/// Ingest → minimize → reassemble round-trip. The most direct
-/// correctness test for `Sedimentree::minimize` in the Automerge use
-/// case: `minimize` must never drop a fragment whose blob data isn't
-/// physically contained in some kept fragment's blob.
-fn ingest_minimize_roundtrip(name: &str, bytes: &[u8]) {
-    let doc = Automerge::load(bytes).expect(name);
-    let original_heads = doc.get_heads();
-    let original_count = doc.get_changes(&[]).len();
-
-    let sed_id = sed_id(bytes);
-    let result = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-    let minimized = result.sedimentree.minimize(&CountLeadingZeroBytes);
-
-    // `minimize` only prunes Sedimentree entries; blob storage is
-    // unaffected, so all blobs collected during ingest are still
-    // addressable.
-    let blob_by_digest: Map<Digest<Blob>, &[u8]> = result
-        .blobs
-        .iter()
-        .map(|blob| (Digest::hash(blob), blob.as_slice()))
-        .collect();
-
-    let kept_fragments: Vec<_> = minimized.fragments().collect();
-    let kept_loose: Vec<_> = minimized.loose_commits().collect();
-
-    // Every kept item must reference a blob we have. A miss means
-    // `minimize` orphaned data.
-    for f in &kept_fragments {
-        let d = f.summary().blob_meta().digest();
-        assert!(
-            blob_by_digest.contains_key(&d),
-            "{name}: minimized fragment references blob {d:?} we don't have",
-        );
-    }
-    for c in &kept_loose {
-        let d = c.blob_meta().digest();
-        assert!(
-            blob_by_digest.contains_key(&d),
-            "{name}: minimized loose commit references blob {d:?} we don't have",
-        );
-    }
-
-    let order = minimized
-        .topsorted_blob_order()
-        .expect("no cycles in minimized tree");
-
-    let mut buf = Vec::new();
-    for item in &order {
-        let digest = match item {
-            SedimentreeItem::Fragment(i) => kept_fragments[*i].summary().blob_meta().digest(),
-            SedimentreeItem::LooseCommit(i) => kept_loose[*i].blob_meta().digest(),
-        };
-        let raw = blob_by_digest
-            .get(&digest)
-            .unwrap_or_else(|| panic!("{name}: blob not found for {item:?}"));
-        buf.extend_from_slice(raw);
-    }
-
-    let mut rebuilt = Automerge::new();
-    rebuilt
-        .load_incremental(&buf)
-        .expect("load minimized blobs");
-
-    assert_eq!(
-        rebuilt.get_heads(),
-        original_heads,
-        "{name}: heads diverged after minimize round-trip \
-         (ingest produced {} fragments / {} loose; \
-          after minimize {} fragments / {} loose)",
-        result.fragment_count,
-        result.loose_count,
-        kept_fragments.len(),
-        kept_loose.len(),
-    );
-    assert_eq!(
-        rebuilt.get_changes(&[]).len(),
-        original_count,
-        "{name}: change count diverged after minimize round-trip \
-         (original={original_count}, rebuilt={})",
-        rebuilt.get_changes(&[]).len(),
-    );
-
-    let orig_keys: Vec<_> = doc.keys(&ROOT).collect();
-    let rebuilt_keys: Vec<_> = rebuilt.keys(&ROOT).collect();
-    assert_eq!(
-        orig_keys, rebuilt_keys,
-        "{name}: root keys diverged after minimize round-trip",
-    );
-}
-
-#[test]
-fn ingest_minimize_roundtrip_s1() {
-    ingest_minimize_roundtrip("S1", include_bytes!("../test-vectors/S1.am"));
-}
-
-#[test]
-fn ingest_minimize_roundtrip_s2() {
-    ingest_minimize_roundtrip("S2", include_bytes!("../test-vectors/S2.am"));
-}
-
-#[test]
-fn ingest_minimize_roundtrip_s3() {
-    ingest_minimize_roundtrip("S3", include_bytes!("../test-vectors/S3.am"));
-}
-
-// NOTE: ingest_minimize_roundtrip_{a1,a2,c1,c2} previously asserted that the
-// concurrent-DAG vectors A1/A2/C1/C2 round-trip through ingest → minimize →
-// blob-concat → Automerge::load_incremental with identical heads. With
-// fragmentation delegated to upstream `Automerge::fragments` /
-// `bundle_fragments`, concurrent siblings can legitimately share members
-// across cached fragments, which breaks the naive blob-concat reassembly
-// these tests performed. The roundtrip invariant for these vectors is now
-// owned by upstream and tracked through `bundle_fragments` correctness
-// there, not duplicated here. The S1/S2/S3 vectors (flat-loose, no cached
-// fragments) remain covered above.
-
-/// Pins (`change_count`, `fragment_count`, `uncovered_count`) for each
-/// egwalker vector. Drift signals a behavioural change in
-/// `ingest_automerge` or `build_fragment_store`/`CountLeadingZeroBytes`.
-/// Update the table when shifting numbers intentionally.
-#[test]
-#[cfg_attr(debug_assertions, ignore = "needs get_changes; release-only")]
-fn egwalker_vector_snapshots() {
-    /// `(name, bytes, expected_changes, expected_fragments, expected_uncovered)`
-    const SNAPSHOTS: &[(&str, &[u8], usize, usize, usize)] = &[
-        ("S1", include_bytes!("../test-vectors/S1.am"), 2, 0, 2),
-        ("S2", include_bytes!("../test-vectors/S2.am"), 2, 0, 2),
-        ("S3", include_bytes!("../test-vectors/S3.am"), 2, 0, 2),
-        ("A1", include_bytes!("../test-vectors/A1.am"), 956, 7, 184),
-        ("A2", include_bytes!("../test-vectors/A2.am"), 3_208, 8, 228),
-        (
-            "C1",
-            include_bytes!("../test-vectors/C1.am"),
-            93_152,
-            380,
-            564,
-        ),
-        (
-            "C2",
-            include_bytes!("../test-vectors/C2.am"),
-            134_477,
-            542,
-            0,
-        ),
-    ];
-
-    for (name, bytes, exp_changes, exp_fragments, exp_uncovered) in SNAPSHOTS {
-        let doc = Automerge::load(bytes).expect(name);
-        let d = decompose_meta(&doc);
-        assert_eq!(
-            d.change_count, *exp_changes,
-            "{name}: change_count drift (got {} expected {exp_changes})",
-            d.change_count
-        );
-        assert_eq!(
-            d.fragment_count, *exp_fragments,
-            "{name}: fragment_count drift (got {} expected {exp_fragments})",
-            d.fragment_count
-        );
-        assert_eq!(
-            d.uncovered_count, *exp_uncovered,
-            "{name}: uncovered_count drift (got {} expected {exp_uncovered})",
-            d.uncovered_count
-        );
-    }
-}
-
-/// Pins `MinimalTreeHash` for each egwalker vector after ingest +
-/// minimize. The strongest sync-correctness regression test:
-/// `MinimalTreeHash` is the canonical content-addressed identity, so
-/// two peers with identical content must compute identical values.
-/// Also catches determinism regressions, since the hash depends on
-/// `minimize`'s output. Update the table when intentional behaviour
-/// changes shift these.
-#[test]
-#[cfg_attr(debug_assertions, ignore = "needs get_changes; release-only")]
-fn egwalker_minimal_hash_snapshots() {
-    use sedimentree_core::sedimentree::MinimalTreeHash;
-
-    /// `(name, bytes, expected_hash_hex)` — pinned after the bundle-format
-    /// rework: per-column DEFLATE in `Bundle`, topo-sorted bundle members
-    /// (small `external` deps list), ingest routed through
-    /// `Automerge::fragments()` instead of `build_fragment_store`.
-    /// The four high-concurrency vectors (A1, A2, C1, C2) re-snapshot
-    /// because fragment blob bytes changed; S1-S3 are unchanged because
-    /// they have no cached fragments and the loose-commit hash structure
-    /// is independent of blob byte format.
-    const SNAPSHOTS: &[(&str, &[u8], &str)] = &[
-        (
-            "S1",
-            include_bytes!("../test-vectors/S1.am"),
-            "eb4edf5719382f067537124deec2696937125cb11ba0fd7f4e54d9abf36282ad",
-        ),
-        (
-            "S2",
-            include_bytes!("../test-vectors/S2.am"),
-            "455c9e3a257f1eae359403408e22347c6856ae60fe00262ddc7a8e7da5c619bd",
-        ),
-        (
-            "S3",
-            include_bytes!("../test-vectors/S3.am"),
-            "ea65d7e84b1b1cc0b0a536ab4d02cbedace6d21aaa033c626ad768fad275bc6b",
-        ),
-        (
-            "A1",
-            include_bytes!("../test-vectors/A1.am"),
-            "4098ca35d635e122eb1def270ea61e958dd321e093131c9c1a5a7389c2c2d099",
-        ),
-        (
-            "A2",
-            include_bytes!("../test-vectors/A2.am"),
-            "53e6f21b41927101d7b1d669e6c1bc569dd9a7a96d7e53e45cf6ad804cd5438d",
-        ),
-        (
-            "C1",
-            include_bytes!("../test-vectors/C1.am"),
-            "cd8276131bcd6fb67e6763c090cf3a53bd5fb5e43765526d8c17dbe7bbaf1b7b",
-        ),
-        (
-            "C2",
-            include_bytes!("../test-vectors/C2.am"),
-            "24b901c5dcf29f37ab66118180cda3233613547f3c65af22820c3742307b872e",
-        ),
-    ];
-
-    for (name, bytes, expected_hex) in SNAPSHOTS {
-        let doc = Automerge::load(bytes).expect(name);
-        let sed_id = sed_id(bytes);
-        let result = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-
-        // Determinism within a single ingest result.
-        let h1: MinimalTreeHash = result.sedimentree.minimal_hash(&CountLeadingZeroBytes);
-        let h2: MinimalTreeHash = result.sedimentree.minimal_hash(&CountLeadingZeroBytes);
-        assert_eq!(
-            h1.as_bytes(),
-            h2.as_bytes(),
-            "{name}: minimal_hash differs across consecutive calls within one process \
-             — DETERMINISM REGRESSION",
-        );
-
-        // Determinism across re-ingest of the same document bytes.
-        let result2 = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-        let h3: MinimalTreeHash = result2.sedimentree.minimal_hash(&CountLeadingZeroBytes);
-        assert_eq!(
-            h1.as_bytes(),
-            h3.as_bytes(),
-            "{name}: minimal_hash differs across re-ingest in one process \
-             — DETERMINISM REGRESSION",
-        );
-
-        let mut hex = String::with_capacity(64);
-        for b in h1.as_bytes() {
-            use core::fmt::Write;
-            write!(&mut hex, "{b:02x}").unwrap();
-        }
-        assert_eq!(
-            hex, *expected_hex,
-            "{name}: minimal_hash drift from pinned snapshot. \
-             If this is intentional, update the SNAPSHOTS table and explain why."
-        );
-    }
-}
-
-/// Minimize twice, then round-trip. Pins idempotence at the Automerge
-/// level — regression coverage for the determinism fix.
-#[test]
-fn ingest_double_minimize_roundtrip_s1() {
-    let bytes = include_bytes!("../test-vectors/S1.am");
-    let doc = Automerge::load(bytes).expect("S1");
-    let original_heads = doc.get_heads();
-
-    let sed_id = sed_id(bytes);
-    let result = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-
-    let once = result.sedimentree.minimize(&CountLeadingZeroBytes);
-    let twice = once.minimize(&CountLeadingZeroBytes);
-    assert_eq!(
-        once, twice,
-        "minimize is not idempotent on real Automerge document",
-    );
-
-    let blob_by_digest: Map<Digest<Blob>, &[u8]> = result
-        .blobs
-        .iter()
-        .map(|blob| (Digest::hash(blob), blob.as_slice()))
-        .collect();
-
-    let kept_fragments: Vec<_> = twice.fragments().collect();
-    let kept_loose: Vec<_> = twice.loose_commits().collect();
-
-    let order = twice.topsorted_blob_order().expect("no cycles");
-    let mut buf = Vec::new();
-    for item in &order {
-        let digest = match item {
-            SedimentreeItem::Fragment(i) => kept_fragments[*i].summary().blob_meta().digest(),
-            SedimentreeItem::LooseCommit(i) => kept_loose[*i].blob_meta().digest(),
-        };
-        let raw = blob_by_digest
-            .get(&digest)
-            .expect("blob not found after double minimize");
-        buf.extend_from_slice(raw);
-    }
-
-    let mut rebuilt = Automerge::new();
-    rebuilt
-        .load_incremental(&buf)
-        .expect("load double-minimized");
-    assert_eq!(
-        rebuilt.get_heads(),
-        original_heads,
-        "heads diverged after double minimize"
-    );
-}
-
-/// Ingest A2 (3,208 changes) via the production `ingest_automerge` API
-/// and verify the roundtrip: topsorted reassembly must produce the same
-/// document heads and change count.
-#[test]
-#[cfg_attr(debug_assertions, ignore = "too slow in debug mode")]
-fn ingest_roundtrip_a2() {
-    let bytes = include_bytes!("../test-vectors/A2.am");
-    let doc = Automerge::load(bytes).expect("A2");
-    let original_heads = doc.get_heads();
-    let original_count = doc.get_changes(&[]).len();
-
-    let sed_id = sed_id(bytes);
-    let result = automerge_sedimentree::ingest::ingest_automerge(&doc, sed_id);
-
-    // Verify ingest accounting: every change is either covered by a
-    // fragment or emitted as a loose commit, with no overlap.
-    assert_eq!(
-        result.covered_count + result.loose_count,
-        original_count,
-        "ingest must account for all changes (covered={}, loose={}, total={})",
-        result.covered_count,
-        result.loose_count,
-        original_count,
-    );
-
-    // Reassemble via topsort
-    let order = result
-        .sedimentree
-        .topsorted_blob_order()
-        .expect("no cycles");
-    let fragments: Vec<_> = result.sedimentree.fragments().collect();
-    let loose: Vec<_> = result.sedimentree.loose_commits().collect();
-
-    let blob_by_digest: Map<Digest<Blob>, &[u8]> = result
-        .blobs
-        .iter()
-        .map(|blob| (Digest::hash(blob), blob.as_slice()))
-        .collect();
-
-    let mut buf = Vec::new();
-    for item in &order {
-        let digest = match item {
-            SedimentreeItem::Fragment(i) => fragments[*i].summary().blob_meta().digest(),
-            SedimentreeItem::LooseCommit(i) => loose[*i].blob_meta().digest(),
-        };
-        let raw = blob_by_digest
-            .get(&digest)
-            .expect("blob not found for item");
-        buf.extend_from_slice(raw);
-    }
-
-    let mut rebuilt = Automerge::new();
-    rebuilt.load_incremental(&buf).expect("load topsorted");
-
-    assert_eq!(rebuilt.get_heads(), original_heads, "heads must match");
-    assert_eq!(
-        rebuilt.get_changes(&[]).len(),
-        original_count,
-        "change count must match"
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Real-world document tests (226k changes, ~6MB)
-//
-// The test file `real-world.am` is not checked in — it contains real user
-// data and is too large for the repository. Tests skip gracefully when the
-// file is absent (the CI egwalker vectors A1/A2/C1/C2/S1-S3 still run).
-//
-// To run these tests locally:
-//
-//   1. Download the document from the shared team drive or regenerate it
-//      from the canonical source. The expected file is the Automerge
-//      binary format (`.am`), ~6 MB, ~226k changes.
-//
-//   2. Place it at:
-//
-//        automerge_sedimentree/test-vectors/real-world.am
-//
-//   3. Run:
-//
-//        cargo test --package automerge_sedimentree --release real_world
-//
-//      The `real_world_metadata_decomposition` test runs in debug mode
-//      too (metadata-only, no `get_changes`). The `real_world_roundtrip`
-//      test requires release mode (~9 seconds).
-//
-//   4. To run *all* roundtrip tests (including the egwalker vectors and
-//      the real-world doc):
-//
-//        cargo test --package automerge_sedimentree --release roundtrip
-//
-// The file is gitignored. Do not check it in.
-// ---------------------------------------------------------------------------
-
-const REAL_WORLD_PATH: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/test-vectors/real-world.am");
-
-fn load_real_world() -> Option<Vec<u8>> {
-    std::fs::read(REAL_WORLD_PATH).ok()
-}
-
-/// Metadata-only decomposition on the 226k-change real-world doc.
-/// Verifies `build_fragment_store` completes and coverage is correct.
-#[test]
-fn real_world_metadata_decomposition() {
-    let Some(bytes) = load_real_world() else {
-        eprintln!("skipping: real-world.am not found at {REAL_WORLD_PATH}");
-        return;
-    };
-
-    let doc = Automerge::load(&bytes).expect("load real-world doc");
-    let d = decompose_meta(&doc);
-
-    eprintln!(
-        "real-world: {} changes, {} fragments, {} uncovered, {} heads",
-        d.change_count,
-        d.fragment_count,
-        d.uncovered_count,
-        d.heads.len(),
-    );
-
-    assert!(d.change_count > 200_000, "expected 200k+ changes");
-    assert!(d.fragment_count > 0, "should produce fragments");
-    assert_eq!(
-        d.covered.len() + d.uncovered_count,
-        d.change_count,
-        "coverage mismatch"
-    );
-}
-
-/// Full bundle roundtrip on the real-world doc.
-/// Only runs in release mode, and only when the file is present.
-#[test]
-#[cfg_attr(debug_assertions, ignore = "too slow in debug mode")]
-fn real_world_roundtrip() {
-    let Some(bytes) = load_real_world() else {
-        eprintln!("skipping: real-world.am not found at {REAL_WORLD_PATH}");
-        return;
-    };
-
-    roundtrip_full("real-world", &bytes);
-}
