+++ automerge_subduction_ingest/src/main.rs
@@ -13,20 +13,93 @@
 //!   document.am
 //! ```
 
-use std::{path::PathBuf, sync::Arc};
+use std::{collections::BTreeSet, path::PathBuf, sync::Arc};
 
 use automerge::Automerge;
-use automerge_sedimentree::ingest::{IngestResult, ingest_automerge_par};
 use clap::Parser;
 use eyre::{Result, WrapErr, eyre};
 use future_form::Sendable;
-use sedimentree_core::id::SedimentreeId;
+use sedimentree_core::{
+    blob::{Blob, BlobMeta},
+    collections::Set,
+    fragment::Fragment,
+    id::SedimentreeId,
+    loose_commit::{LooseCommit, id::CommitId},
+    sedimentree::Sedimentree,
+};
 use subduction_core::{
     handshake::audience::Audience, policy::open::OpenPolicy, storage::memory::MemoryStorage,
     subduction::builder::SubductionBuilder, transport::message::MessageTransport,
 };
 use subduction_crypto::signer::memory::MemorySigner;
 use subduction_websocket::tokio::{TimeoutTokio, TokioSpawn, client::TokioWebSocketClient};
+
+/// Result of ingesting an Automerge document.
+struct IngestResult {
+    sedimentree: Sedimentree,
+    blobs: Vec<Blob>,
+    change_count: usize,
+    covered_count: usize,
+    loose_count: usize,
+    fragment_count: usize,
+}
+
+/// Ingest an Automerge document into a [`Sedimentree`].
+///
+/// Thin adapter over [`Automerge::fragments`] and
+/// [`Automerge::bundle_fragments`]: maps each level-1+ `automerge::Fragment`
+/// into a sedimentree [`Fragment`] and each level-0 fragment into a
+/// [`LooseCommit`].
+fn ingest_automerge(doc: &Automerge, sedimentree_id: SedimentreeId) -> IngestResult {
+    let cached = doc.fragments(1..);
+    let loose = doc.fragments(0..=0);
+    let cached_bytes = doc.bundle_fragments(cached.iter().cloned());
+    let loose_bytes = doc.bundle_fragments(loose.iter().cloned());
+
+    let mut fragments = Vec::with_capacity(cached.len());
+    let mut blobs = Vec::with_capacity(cached.len() + loose.len());
+    let mut covered: Set<CommitId> = Set::new();
+    for (f, raw) in cached.iter().zip(cached_bytes) {
+        for m in &f.members {
+            covered.insert(CommitId::new(m.0));
+        }
+        let blob = Blob::new(raw);
+        let boundary: BTreeSet<CommitId> = f.boundary.iter().map(|h| CommitId::new(h.0)).collect();
+        let checkpoints: Vec<CommitId> = f.checkpoints.iter().map(|h| CommitId::new(h.0)).collect();
+        let meta = BlobMeta::new(&blob);
+        fragments.push(Fragment::new(
+            sedimentree_id,
+            CommitId::new(f.head.0),
+            boundary,
+            &checkpoints,
+            meta,
+        ));
+        blobs.push(blob);
+    }
+
+    let mut loose_commits = Vec::with_capacity(loose.len());
+    for (f, raw) in loose.iter().zip(loose_bytes) {
+        let head = CommitId::new(f.head.0);
+        let parents: BTreeSet<CommitId> = f.boundary.iter().map(|p| CommitId::new(p.0)).collect();
+        let blob = Blob::new(raw);
+        let meta = BlobMeta::new(&blob);
+        loose_commits.push(LooseCommit::new(sedimentree_id, head, parents, meta));
+        blobs.push(blob);
+    }
+
+    let fragment_count = fragments.len();
+    let loose_count = loose_commits.len();
+    let covered_count = covered.len();
+
+    IngestResult {
+        sedimentree: Sedimentree::new(fragments, loose_commits),
+        blobs,
+        change_count: doc.get_changes_meta(&[]).len(),
+        covered_count,
+        loose_count,
+        fragment_count,
+    }
+}
 
 /// Ingest an Automerge document into a Subduction sync server.
 #[derive(Debug, Parser)]
@@ -229,7 +302,7 @@ async fn main() -> Result<()> {
 
     // Ingest: automerge → sedimentree.
     eprintln!("ingesting...");
-    let result = ingest_automerge_par(&doc, sed_id);
+    let result = ingest_automerge(&doc, sed_id);
     print_ingest_stats(&result, sed_id);
 
     if args.dry_run {
